ai-engine/requirements.txt
httpx
redis
google-genai
python-dotenv
fastapi
uvicorn
MetaTrader5
rich
psycopg[binary]
